obsługa AutoCompleteTextView

+++ app/src/main/java/com/example/myapplication/Database/CatsHeathBookOpenHelper.java
@@ -8,9 +8,6 @@ import android.database.sqlite.SQLiteOpenHelper;
 import android.widget.Toast;
 
 import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-import com.example.myapplication.Fragments.DodanieFragment;
 
 import static com.example.myapplication.Database.CatsHealthBookDatabseContract.KalendarzEntry.COLUMN_DATA;
 import static com.example.myapplication.Database.CatsHealthBookDatabseContract.SpisChorobEntry.COLUMN_NAZWA_CHOROBY;
@@ -21,9 +18,13 @@ import static com.example.myapplication.Database.CatsHealthBookDatabseContract.S
 public class CatsHeathBookOpenHelper extends SQLiteOpenHelper {
     private SQLiteDatabase db;
     private Context context;
-    private Fragment fragment;
     private static final String DATABSE_NAME = "CatsHealthBook.db"; //nazwa aszego pliku, w którym będzie baza.
     private static final int DATABASE_VERSION = 1;
+    protected Cursor cursor;
+
+    public Cursor getCursor() {
+        return cursor;
+    }
 
     public CatsHeathBookOpenHelper(@Nullable Context context) {
         super(context, DATABSE_NAME, null, DATABASE_VERSION);
@@ -42,16 +43,27 @@ public class CatsHeathBookOpenHelper extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + CatsHealthBookDatabseContract.SpisZwierzatEntry.TABLE_NAME);
-        db.execSQL("DROP TABLE IF EXISTS " + CatsHealthBookDatabseContract.KalendarzEntry.TABLE_NAME);
-        db.execSQL("DROP TABLE IF EXISTS " + CatsHealthBookDatabseContract.SpisZabiegowEntry.TABLE_NAME);
-        db.execSQL("DROP TABLE IF EXISTS " + CatsHealthBookDatabseContract.SpisLekowEntry.TABLE_NAME);
-        db.execSQL("DROP TABLE IF EXISTS " + CatsHealthBookDatabseContract.SpisChorobEntry.TABLE_NAME);
-        onCreate(db);
+        if (newVersion > oldVersion) {
+            db.execSQL("DROP TABLE IF EXISTS " + CatsHealthBookDatabseContract.SpisZwierzatEntry.TABLE_NAME);
+            db.execSQL("DROP TABLE IF EXISTS " + CatsHealthBookDatabseContract.KalendarzEntry.TABLE_NAME);
+            db.execSQL("DROP TABLE IF EXISTS " + CatsHealthBookDatabseContract.SpisZabiegowEntry.TABLE_NAME);
+            db.execSQL("DROP TABLE IF EXISTS " + CatsHealthBookDatabseContract.SpisLekowEntry.TABLE_NAME);
+            db.execSQL("DROP TABLE IF EXISTS " + CatsHealthBookDatabseContract.SpisChorobEntry.TABLE_NAME);
+            onCreate(db);
+        }
     }
 
+    public void saveToDatabase() {
+        db = this.getWritableDatabase();
+    }
+
+    public void readFromDatabase() {
+        db = this.getReadableDatabase();
+    }
+
+
     public void addMed(String nazwa_leku, String dodatkowe_informacje) {
-        SQLiteDatabase db = this.getWritableDatabase();
+        saveToDatabase();
         ContentValues cv = new ContentValues();
 
         cv.put(COLUMN_NAZWA_LEKU, nazwa_leku);
@@ -68,7 +80,7 @@ public class CatsHeathBookOpenHelper extends SQLiteOpenHelper {
     ;
 
     public void addDisease(String nazwa_choroby, String dodatkowe_informacje) {
-        SQLiteDatabase db = this.getWritableDatabase();
+        saveToDatabase();
         ContentValues cv = new ContentValues();
 
         cv.put(COLUMN_NAZWA_CHOROBY, nazwa_choroby);
@@ -83,7 +95,7 @@ public class CatsHeathBookOpenHelper extends SQLiteOpenHelper {
 
 
     public void addTreatment(String nazwa_zabiegu, String dodatkowe_informacje) {
-        SQLiteDatabase db = this.getWritableDatabase();
+        saveToDatabase();
         ContentValues cv = new ContentValues();
 
         cv.put(COLUMN_NAZWA_ZABIEGU, nazwa_zabiegu);
@@ -96,10 +108,10 @@ public class CatsHeathBookOpenHelper extends SQLiteOpenHelper {
         }
     }
 
-    public void addCat(String nazwa_zabiegu, String nazwa_leku, String nazwa_choroby, String data, String rok_urodzenia) {
-        SQLiteDatabase db = this.getWritableDatabase();
+    public void addCat(String imie_kota, String nazwa_zabiegu, String nazwa_leku, String nazwa_choroby, String data, String rok_urodzenia) {
+        saveToDatabase();
         ContentValues cv = new ContentValues();
-
+        cv.put(CatsHealthBookDatabseContract.SpisZwierzatEntry.COLUMN_IMIE_KOTA, imie_kota);
         cv.put(CatsHealthBookDatabseContract.SpisZwierzatEntry.COLUMN_NAZWA_ZABIEGU, nazwa_zabiegu);
         cv.put(CatsHealthBookDatabseContract.SpisZwierzatEntry.COLUMN_NAZWA_LEKU, nazwa_leku);
         cv.put(CatsHealthBookDatabseContract.SpisZwierzatEntry.COLUMN_NAZWA_CHOROBY, nazwa_choroby);
@@ -114,7 +126,7 @@ public class CatsHeathBookOpenHelper extends SQLiteOpenHelper {
     }
 
     public void addDate(String data, String dodatkowe_informacje) {
-        SQLiteDatabase db = this.getWritableDatabase();
+        saveToDatabase();
         ContentValues cv = new ContentValues();
 
         cv.put(COLUMN_DATA, data);
@@ -129,9 +141,8 @@ public class CatsHeathBookOpenHelper extends SQLiteOpenHelper {
 
 
     public Cursor readAllDataMeds() {
-        SQLiteDatabase db = this.getReadableDatabase();
+        readFromDatabase();
         String query = "SELECT * FROM " + CatsHealthBookDatabseContract.SpisLekowEntry.TABLE_NAME;
-        Cursor cursor = null;
         if (db != null) {
             cursor = db.rawQuery(query, null);
         }
@@ -141,9 +152,8 @@ public class CatsHeathBookOpenHelper extends SQLiteOpenHelper {
     }
 
     public Cursor readAllDataTreatments() {
-        SQLiteDatabase db = this.getReadableDatabase();
+        readFromDatabase();
         String query = "SELECT * FROM " + CatsHealthBookDatabseContract.SpisZabiegowEntry.TABLE_NAME;
-        Cursor cursor = null;
         if (db != null) {
             cursor = db.rawQuery(query, null);
         }
@@ -151,13 +161,55 @@ public class CatsHeathBookOpenHelper extends SQLiteOpenHelper {
     }
 
     public Cursor readAllDataDiseases() {
-        SQLiteDatabase db = this.getReadableDatabase();
+        readFromDatabase();
         String query = "SELECT * FROM " + CatsHealthBookDatabseContract.SpisChorobEntry.TABLE_NAME;
-        Cursor cursor = null;
         if (db != null) {
             cursor = db.rawQuery(query, null);
         }
         return cursor;
     }
-}
 
+    public Cursor readFromDatabaseOnlyNazwaLeku() {
+        readFromDatabase();
+        String query = "SELECT " + CatsHealthBookDatabseContract.SpisLekowEntry.COLUMN_NAZWA_LEKU + " FROM " + CatsHealthBookDatabseContract.SpisLekowEntry.TABLE_NAME;
+        if (db != null) {
+            cursor = db.rawQuery(query, null);
+        }
+        cursor.moveToFirst();
+        return cursor;
+
+    }
+
+    public Cursor readFromDatabaseOnlyNazwaChoroby() {
+        readFromDatabase();
+        String query = "SELECT " + CatsHealthBookDatabseContract.SpisChorobEntry.COLUMN_NAZWA_CHOROBY + " FROM " + CatsHealthBookDatabseContract.SpisChorobEntry.TABLE_NAME;
+        if (db != null) {
+            cursor = db.rawQuery(query, null);
+        }
+        cursor.moveToFirst();
+        return cursor;
+    }
+
+    public Cursor readFromDatabaseOnlyNazwaZabiegu() {
+        readFromDatabase();
+        String query = "SELECT " + CatsHealthBookDatabseContract.SpisZabiegowEntry.COLUMN_NAZWA_ZABIEGU + " FROM " + CatsHealthBookDatabseContract.SpisZabiegowEntry.TABLE_NAME;
+        if (db != null) //jezeli baza danych nie jest pusta to wykonaj query
+             {
+            cursor = db.rawQuery(query, null);
+        }
+        cursor.moveToFirst();
+        return cursor;
+        //kurson przechowuje nasze query
+    }
+
+    public Cursor readFromDatabaseOnlyImieKota(){
+        readFromDatabase();
+        String query = "SELECT " + CatsHealthBookDatabseContract.SpisZwierzatEntry.COLUMN_IMIE_KOTA + " FROM " + CatsHealthBookDatabseContract.SpisZwierzatEntry.TABLE_NAME;
+        if(db != null){
+            cursor = db.rawQuery(query, null);
+        }
+        cursor.moveToFirst();
+        return cursor;
+    }
+
+}
